refactor workflow file
